requirements.txt
flask
azure-ai-textanalytics
gunicorn
